extract the platform API into a separate header
